@@ -23,7 +23,7 @@ setuptools.setup(
 
     # pip 9.0+ will inspect this field when installing to help users install a
     # compatible version of the library for their Python version.
-    python_requires='>=3.5',
+    python_requires='>=3.6',
 
     # There are some peculiarities on how to include package data for source
     # distributions using setuptools. You also need to add entries for package
@@ -50,7 +50,6 @@ setuptools.setup(
         'Programming Language :: Python',
         'Programming Language :: Python :: 3',
         'Programming Language :: Python :: 3 :: Only',
-        'Programming Language :: Python :: 3.5',
         'Programming Language :: Python :: 3.6',
         'Programming Language :: Python :: 3.7',
         'Programming Language :: Python :: 3.8',
